refactor: Split useVercel hooks into two hooks
